+++ packages/expo-audio/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### 🐛 Bug fixes
 
+- [iOS] Move the blocking `AVAudioSession.setActive(false)` deactivation off the main thread and skip stale deactivations to fix app hangs during frequent back-to-back playback. ([#47079](https://github.com/expo/expo/pull/47079) by [@sbs44](https://github.com/sbs44))
 - [Android] Fix stale lock screen artwork when updating metadata without an `artworkUrl`. ([#45738](https://github.com/expo/expo/pull/45738) by [@behenate](https://github.com/behenate))
 
 ### 💡 Others